pyproject.toml
[tool.poetry]
name = "graduation-project"
version = "0.1.0"
description = "graduation project on automation in python"
authors = ["MozzhukhinRA <roma.moz@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
allure-pytest = "==2.15.0"
allure-python-commons = "==2.15.0"
pytest = "==8.4.1"
python-dotenv = "==1.1.1"
requests = "2.32.4"
selene = "==2.0.0rc9"
selenium = "==4.34.2"
appium-python-client = "~=5.2.4"
config = "~=0.5.1"
faker = "^37.12.0"
pydantic = "^2.12.4"
jsonschema = "^4.25.1"


[tool.poetry.group.dev.dependencies]
ruff = "^0.14.8"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

